fix(review): move hook diagnostic infra to hook_registry, fix scope bug
Addresses four review findings:
- server/helpers.py imported from cli/ (L3 peer) — moved
  _claude_settings_path, _load_settings_data, _extract_cmds,
  HookDriftResult, _count_hook_registry_drift, find_broken_hook_scripts
  to hook_registry.py (package root, accessible by all layers)
- _build_hook_diagnostic_warning used 'project' scope but hooks deploy
  to 'user' scope — fixed to 'user'
- _check_hook_registry_drift duplicated I/O by re-calling
  generate_hooks_json() and _load_settings_data() — HookDriftResult now
  includes orphaned_cmds frozenset, eliminating the second call
- _extract_cmds promoted but not reused — now shared via hook_registry
